Revisions based on review comments
