--- tests/integration/security/plugin_ca_key_cert/main_test.go
@@ -1,61 +1,0 @@
-// Copyright 2020 Istio Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package plugincakeycert
-
-import (
-	"testing"
-
-	"istio.io/istio/tests/integration/security/util/cert"
-
-	"istio.io/istio/pkg/test/framework"
-	"istio.io/istio/pkg/test/framework/components/galley"
-	"istio.io/istio/pkg/test/framework/components/istio"
-	"istio.io/istio/pkg/test/framework/components/pilot"
-	"istio.io/istio/pkg/test/framework/label"
-	"istio.io/istio/pkg/test/framework/resource"
-	"istio.io/istio/pkg/test/framework/resource/environment"
-)
-
-var (
-	inst istio.Instance
-	g    galley.Instance
-	p    pilot.Instance
-)
-
-func TestMain(m *testing.M) {
-	// Integration test for plugging in the CA key and certificate.
-	// Istio CA will use the plugged in CA certificate to issue certificate
-	// and the CA certificate is mounted to a configmap in each namespace
-	// for data plane to control plane TLS authentication.
-	framework.
-		NewSuite("plugin_ca_key_cert_test", m).
-		// k8s is required because the plugin CA key and certificate are stored in a k8s secret.
-		RequireEnvironment(environment.Kube).
-		RequireSingleCluster().
-		Label(label.CustomSetup).
-		SetupOnEnv(environment.Kube, istio.Setup(&inst, nil, cert.CreateCASecret)).
-		Setup(func(ctx resource.Context) (err error) {
-			if g, err = galley.New(ctx, galley.Config{}); err != nil {
-				return err
-			}
-			if p, err = pilot.New(ctx, pilot.Config{
-				Galley: g,
-			}); err != nil {
-				return err
-			}
-			return nil
-		}).
-		Run()
-}
